pyproject.toml
[project]
name = "rightboard"
version = "0.1.0"
description = "Python helper scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "pyserial>=3.5",
]
